+++ CHANGELOG.md
@@ -32,11 +32,13 @@
 
 ### Additions
 - Opening doors with dodgerolls will now deal high-knockback damage to entities.
+- Added velocity-based tilting (rotation offset) effects to enemies and NPCs, previously only seen on player characters. Toggled by the `Visuals.EnableEnemyTiltingEffects` config entry.
 
 ### Changes
 - Heavily improved combat info tooltips. They now have far more information and feature formatting and color highlighting. They were also added to Hammers, Axes and Pickaxes.
 - Changed mana pickups' particles, fixed these particles being influenced by the local player's movement.
 - Changed how the `Accessibility.ForceAutoReuse` option works. Now it will only exist to make the vanilla global auto-reuse option that was added in Terraria 1.4.4 be enabled by default for players of Overhaul. This change also removes a 2-tick (33.33 milliseconds) use speed penalty from items that had its auto-reuse (auto-swing) forced on by Overhaul, as rebalancing has now been done in vanilla.
+- Slightly altered velocity-based tilting effects on player characters.
 
 ### Localization
 - Polish - PR [#205](https://github.com/Mirsario/TerrariaOverhaul/pull/205) by **J00niper**.